Refactor BaseController and LibraryRegistryAnnotator into shared file
