+++ tests/conftest.py
@@ -11,6 +11,7 @@ from sqlalchemy.orm.session import Session
 from library_registry.admin import admin
 from library_registry.drm import drm
 from library_registry.library_registration_protocol import libr
+from library_registry.library_list import libr_list
 
 from library_registry.app import create_app
 from library_registry.config import Configuration
@@ -87,6 +88,7 @@ def app(db_session):
     app.register_blueprint(drm)
     app.register_blueprint(admin)
     app.register_blueprint(libr)
+    app.register_blueprint(libr_list)
     yield app
